OneDrive/Desktop/dbms project cp Land rms/requirements.txt
# Core Flask Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
cryptography==41.0.7

# Database Migrations
Flask-Migrate==4.0.5
alembic==1.13.0

# Authentication & Authorization
Flask-Login==0.6.3

# Forms & Validation
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0

# Email Support
Flask-Mail==0.10.0

# PDF Generation
reportlab==4.0.7
WeasyPrint==60.1

# Excel Generation
openpyxl==3.1.2
xlsxwriter==3.1.9

# Image Processing
Pillow==10.4.0
pytesseract==0.3.10  # OCR for document processing

# Charts and Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# GIS and Mapping
folium==0.15.1
geopy==2.4.1

# Advanced Data Processing
pandas==2.1.4
numpy==1.26.2

# QR Code Generation
qrcode==7.4.2

# Task Scheduling
celery==5.3.4
redis==5.0.1

# Rate Limiting
Flask-Limiter==3.5.0

# API Documentation
flasgger==0.9.7.1

# WebSockets for real-time features
Flask-SocketIO==5.3.6
python-socketio==5.10.0

# Advanced Authentication
PyJWT==2.8.0
Flask-JWT-Extended==4.6.0

# Data Validation
marshmallow==3.20.1

# Caching
Flask-Caching==2.1.0

# Environment Variables
python-dotenv==1.0.0

# Date & Time Utilities
python-dateutil==2.8.2

# Production Server
gunicorn==21.2.0

# Security
itsdangerous==2.1.2

# HTTP Requests (for API integration)
requests==2.31.0

# JSON Web Tokens (for API authentication)
PyJWT==2.8.0

# Caching
Flask-Caching==2.1.0

# Rate Limiting
Flask-Limiter==3.5.0

# CORS Support (for API)
Flask-CORS==4.0.0

# Logging
colorlog==6.8.0

# Utilities
python-slugify==8.0.1
